Added Doxygen

+++ Global/math/pow.c
@@ -2,8 +2,8 @@
 
 unsigned long long pow(double x, int y)
 {
-    int ans = 1.0;
-    for (int i = 0; i < y; i++)
+    unsigned long long ans = 1.0;
+    for (unsigned int i = 0; i < y; i++)
         ans *= x;
     return ans;
 }
